requirements.txt
einops==0.2.0
paddlepaddle
pyclipper
lmdb
pandas
numpy==1.23.5
matplotlib
opencv-python
torch
torchvision
gdown
PyYAML
shapely
scikit-image
imgaug
vietocr
python-Levenshtein
setuptools
paddleocr
albumentations